+++ src/main/java/kr/co/air/dtos/SeatDto.java
@@ -1,5 +1,7 @@
 package kr.co.air.dtos;
 
+import java.math.BigDecimal;
+
 import lombok.Data;
 
 @Data
@@ -7,6 +9,6 @@ public class SeatDto {
     private int seatIdx;  
     private int schedIdx;
     private String seatTypeName;   // 좌석 형태명
-    private long seatTypePrice;
+    private BigDecimal seatTypePrice;
     private long seatCount;
 }
